pyproject.toml
[tool.poetry]
name = "tg_blog_client"
version = "0.1.0"
description = ""
authors = ["Denzzelindahouse <deniro-094@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.11"
requests = "^2.27.1"
bs4 = "^0.0.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
